+++ app/Http/Controllers/Admin/HabitatsAdminController.php
@@ -9,35 +9,45 @@ use App\Requests\HabitatsFormRequest;
 use Exception;
 use Illuminate\Contracts\View\View;
 use Illuminate\Http\RedirectResponse;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 
 class HabitatsAdminController extends Controller
 {
-    public function index(): View
+    public function index(): View|RedirectResponse
     {
-        $habitats = Habitat::all();
-        return view('page.habitats', compact('habitats'));
-    }
+        if (Gate::denies('view', Habitat::class))
+        {
+            return redirect()->route('dashboard');
+        }
 
-    public function habitats(): View
-    {
-        return view('page.habitats');
+        $habitats = Habitat::all();
+        return view('admin.zoo.habitats.index', compact('habitats'));
     }
 
     /**
      * @throws Exception
      */
-    public function show(string $name): View
+    public function show(string $name): View|RedirectResponse
     {
+        if (Gate::denies('view', Habitat::class))
+        {
+            return redirect()->route('dashboard');
+        }
         $habitat = Habitat::query()->where('name','=', $name)->with('animals')->first();
         if (!$habitat){
             throw new Exception("Tu n'as pas d'habitat", 404);
         }
-        return view('page.habitat', compact('habitat'));
+        return view('admin.zoo.habitats.show', compact('habitat'));
     }
 
     public function create(HabitatsFormRequest $request): RedirectResponse
     {
+        if (Gate::denies('create', Habitat::class))
+        {
+            return redirect()->route('dashboard');
+        }
+
         $zoo = Zoo::query()->where('name', '=', 'Arcadia')->first();
         $path = $request->file('image')->getClientOriginalName();
 
@@ -58,13 +68,21 @@ class HabitatsAdminController extends Controller
         return redirect()->route('home');
     }
 
-    public function createForm(): View
+    public function createForm(): View|RedirectResponse
     {
+        if (Gate::denies('create', Habitat::class))
+        {
+            return redirect()->route('dashboard');
+        }
         return view('admin.zoo.habitats.create');
     }
 
-    public function edit(string $name): View
+    public function edit(string $name): View|RedirectResponse
     {
+        if (Gate::denies('update', Habitat::class))
+        {
+            return redirect()->route('dashboard');
+        }
         $zoo = Zoo::query()->where('name', '=', 'Arcadia')->first();
         $habitat = Habitat::query()->where('name','=', $name)->first();
 
@@ -73,6 +91,11 @@ class HabitatsAdminController extends Controller
 
     public function update(HabitatsFormRequest $request, string $name): RedirectResponse
     {
+        if (Gate::denies('update', Habitat::class))
+        {
+            return redirect()->route('dashboard');
+        }
+
         $zoo = Zoo::query()->where('name', '=', 'Arcadia')->first();
         $habitat = Habitat::query()->where('name', '=', $name)->first();
         $path = $request->file('image')->getClientOriginalName();
@@ -93,6 +116,10 @@ class HabitatsAdminController extends Controller
 
     public function delete(string $name): RedirectResponse
     {
+        if (Gate::denies('delete', Habitat::class))
+        {
+            return redirect()->route('dashboard');
+        }
         $habitat = Habitat::query()->where('name', '=', $name)->first();
         Storage::disk('public')->delete('asset/images/'.$habitat->image);
         $habitat->delete();
