+++ app/Http/Controllers/Admin/OpeningsAdminController.php
@@ -8,17 +8,28 @@ use App\Models\Zoo;
 use App\Requests\OpeningsFormRequest;
 use Exception;
 use Illuminate\Http\RedirectResponse;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\View\View;
 
 class OpeningsAdminController extends Controller
 {
-    public function index(): View
+    public function index(): View|RedirectResponse
     {
+        if (Gate::denies('view', Opening::class)) {
+            return redirect()->route('dashboard');
+        }
         return view('home');
     }
 
-    public function show(string $day_open): View
+    /**
+     * @throws Exception
+     */
+    public function show(string $day_open): View|RedirectResponse
     {
+        if (Gate::denies('view', Opening::class)) {
+            return redirect()->route('dashboard');
+        }
+
         $openings = Opening::query()->where('day_open', $day_open)->first();
         if (!$openings) {
             throw new Exception("Il n'y a pas d'horaires", 404);
@@ -28,6 +39,10 @@ class OpeningsAdminController extends Controller
 
     public function create(OpeningsFormRequest $request): RedirectResponse
     {
+        if (Gate::denies('create', Opening::class)) {
+            return redirect()->route('dashboard');
+        }
+
         $zoo = Zoo::query()->where('name', '=', 'Arcadia')->first();
 
         $opening = new Opening();
@@ -41,13 +56,21 @@ class OpeningsAdminController extends Controller
         return redirect()->route('home');
     }
 
-    public function createForm(): View
+    public function createForm(): View|RedirectResponse
     {
+        if (Gate::denies('create', Opening::class)) {
+            return redirect()->route('dashboard');
+        }
+
         return view('admin.zoo.openings.create');
     }
 
-    public function edit(string $day_open): View
+    public function edit(string $day_open): View|RedirectResponse
     {
+        if (Gate::denies('update', Opening::class)) {
+            return redirect()->route('dashboard');
+        }
+
         $zoo = Zoo::query()->where('name', '=', 'Arcadia')->first();
         $opening = Opening::query()->where('day_open', '=', $day_open)->first();
 
@@ -56,6 +79,10 @@ class OpeningsAdminController extends Controller
 
     public function update(OpeningsFormRequest $request, string $day_open): RedirectResponse
     {
+        if (Gate::denies('update', Opening::class)) {
+            return redirect()->route('dashboard');
+        }
+
         $zoo = Zoo::query()->where('name', '=', 'Arcadia')->first();
         $opening = Opening::query()->where('day_open', '=', $day_open)->first();
 
@@ -71,6 +98,10 @@ class OpeningsAdminController extends Controller
 
     public function delete(string $day_open): RedirectResponse
     {
+        if (Gate::denies('delete', Opening::class)) {
+            return redirect()->route('dashboard');
+        }
+
         $opening = Opening::query()->where('day_open', '=', $day_open)->first();
         $opening->delete();
         return redirect()->route('home');
